Refactored code, worked on status page
