+++ simplq/src/components/common/InputField.jsx
@@ -1,10 +1,9 @@
-/* eslint-disable import/prefer-default-export */
 import React from 'react';
 import TextField from '@material-ui/core/TextField';
 import { StylesProvider } from '@material-ui/core/styles';
 
 // StylesProvider added to enable CSS use through className prop
-export const InputField = (props) => (
+export default (props) => (
   <StylesProvider injectFirst>
     <TextField
       placeholder={props.placeholder}
